+++ packages/adhere-util-communication-ajax/src/ajax.ts
@@ -4,7 +4,7 @@ import GlobalIndicator from '@baifendian/adhere-ui-globalindicator';
 import Util from '@baifendian/adhere-util';
 import intl from '@baifendian/adhere-util-intl';
 
-import type { IConfig, ISendArg, ISendPrepareArg, Method } from './types';
+import type { IConfig, ISendArg, ISendPrepareArg, Method, Prepare, SendResult } from './types';
 
 // 是否触发过402
 let trigger402 = false;
@@ -98,15 +98,8 @@ class Ajax {
    * @param data
    * @param arg
    */
-  get(
-    this: Ajax,
-    { data, ...arg }: ISendArg,
-  ): {
-    xhr?: XMLHttpRequest | null;
-    contentType?: string | null;
-    promise: Promise<any>;
-  } {
-    let prepare: { xhr?: XMLHttpRequest | null; contentType?: string | null } = {};
+  get(this: Ajax, { data, ...arg }: ISendArg): SendResult {
+    let prepare: Prepare = {};
 
     const promise = new Promise((resolve, reject) => {
       prepare = sendPrepare.call(
@@ -143,14 +136,7 @@ class Ajax {
    * post
    * @param params
    */
-  post(
-    this: Ajax,
-    params: ISendArg,
-  ): {
-    xhr?: XMLHttpRequest | null;
-    contentType?: string | null;
-    promise: Promise<any>;
-  } {
+  post(this: Ajax, params: ISendArg): SendResult {
     return complexRequest.call(this, 'post', params);
   }
 
@@ -158,14 +144,7 @@ class Ajax {
    * path
    * @param params
    */
-  path(
-    this: Ajax,
-    params: ISendArg,
-  ): {
-    xhr?: XMLHttpRequest | null;
-    contentType?: string | null;
-    promise: Promise<any>;
-  } {
+  path(this: Ajax, params: ISendArg): SendResult {
     return complexRequest.call(this, 'path', params);
   }
 
@@ -173,14 +152,7 @@ class Ajax {
    * put
    * @param params
    */
-  put(
-    this: Ajax,
-    params: ISendArg,
-  ): {
-    xhr?: XMLHttpRequest | null;
-    contentType?: string | null;
-    promise: Promise<any>;
-  } {
+  put(this: Ajax, params: ISendArg): SendResult {
     return complexRequest.call(this, 'put', params);
   }
 
@@ -188,14 +160,7 @@ class Ajax {
    * delete
    * @param params
    */
-  delete(
-    this: Ajax,
-    params: ISendArg,
-  ): {
-    xhr?: XMLHttpRequest | null;
-    contentType?: string | null;
-    promise: Promise<any>;
-  } {
+  delete(this: Ajax, params: ISendArg): SendResult {
     return complexRequest.call(this, 'delete', params);
   }
 }
@@ -501,10 +466,7 @@ function sendPrepare(
     ...curConfig // timeout && withCredentials && events
   }: ISendPrepareArg,
   { resolve, reject },
-): {
-  xhr: XMLHttpRequest | null;
-  contentType: string | null;
-} {
+): Prepare {
   let indicator;
 
   const defaultLoadingText = `${intl.v('加载中')}...`;
@@ -628,8 +590,9 @@ function sendPrepare(
  * getSendParams
  * @param data
  * @param contentType
+ * @param customSendJSONStringify
  */
-function getSendParams({ data, contentType = '' }) {
+function getSendParams({ data, contentType = '', customSendJSONStringify }) {
   // 四种Content-Type的处理(也就是send的参数)
 
   // application/json
@@ -648,6 +611,9 @@ function getSendParams({ data, contentType = '' }) {
    */
   if (contentType.startsWith(Ajax.CONTENT_TYPE_APPLICATION_JSON) && Util.isRef(data)) {
     // console.log('数据需要被转换成JSON字符串', JSON.stringify(data));
+    if (customSendJSONStringify) {
+      return JSON.stringify(data, customSendJSONStringify);
+    }
     return JSON.stringify(data);
   }
 
@@ -709,7 +675,12 @@ function getSendParams({ data, contentType = '' }) {
    */
   if (contentType.startsWith(Ajax.CONTENT_TYPE_TEXT_PLAIN)) {
     if (Util.isString(data)) return data;
-    if (Util.isObject(data)) return JSON.stringify(data);
+    if (Util.isObject(data)) {
+      if (customSendJSONStringify) {
+        return JSON.stringify(data, customSendJSONStringify);
+      }
+      return JSON.stringify(data);
+    }
   }
 
   return data.toString();
@@ -720,19 +691,8 @@ function getSendParams({ data, contentType = '' }) {
  * @param method
  * @param params
  */
-function complexRequest(
-  this: Ajax,
-  method: Method,
-  params: ISendArg,
-): {
-  xhr?: XMLHttpRequest | null;
-  contentType?: string | null;
-  promise: Promise<any>;
-} {
-  let prepare: {
-    xhr?: XMLHttpRequest | null;
-    contentType?: string | null;
-  } = {};
+function complexRequest(this: Ajax, method: Method, params: ISendArg): SendResult {
+  let prepare: Prepare = {};
 
   const promise = new Promise((resolve, reject) => {
     prepare = sendPrepare.call(
@@ -759,6 +719,7 @@ function complexRequest(
         getSendParams.call(this, {
           data: params.data,
           contentType: contentType!,
+          customSendJSONStringify: params.customSendJSONStringify,
         }),
       );
     }